Rename files to match type rename
